# restapi-lab3/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "info": "WRC automobili",
        "version": "3.0",
        "contact": "mc52657@fer.hr",
        "licence": "CC Zero v1.0 Universal"
    },
    "paths": {
...
        "/vozaci/": {
            "get": {
                "summary": "Podaci o svim vozacima",
                "description": "Dohvat podataka o svim vozacima",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Podaci o svim vozacima"
                    },
                    "500": {
                        "description": "Greska sa bazom podataka"
                    }
                }
            },
            "post": {
                "summary": "Dodavanje vozaca",
                "description": "Dodavanje novog vozaca po prilozenim podacima",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "id_vozaca",
                        "in": "body",
                        "description": "ID novog vozaca",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "ime_vozaca",
                        "in": "body",
                        "description": "Ime novog vozaca",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "prezime_vozaca",
                        "in": "body",
                        "description": "Prezime novog vozaca",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Dodan vozac"
                    },
                    "400": {
                        "description": "Nedostaju podaci o vozacu"
                    }

                }
            },
            "put": {
                "summary": "Metoda nije implementirana",
                "description": "Metoda nije implementirana",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "501": {
                        "description": "Nije implementirano"
                    }
                }
            },
            "delete": {
                "summary": "Metoda nije implementirana",
                "description": "Metoda nije implementirana",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "501": {
                        "description": "Nije implementirano"
                    }
                }
            }

        },
        "/vozaci/{id}": {
            "get": {
                "summary": "Podaci o vozacu sa trazenim ID-om",
                "description": "Dohvat podataka o vozacu sa trazenim ID-om",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID zeljenog vozaca",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }],
                "responses": {
                    "200": {
                        "description": "Uspjesan dohvat"
                    },
                    "404": {
                        "description": "Ne postoji vozac sa tim ID-om"
                    }
                }
            },
            "post": {
                "summary": "Metoda nije implementirana",
                "description": "Metoda nije implementirana",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "501": {
                        "description": "Nije implementirano"
                    }
                }
            },
            "put": {
                "summary": "Azuriranje podataka o vozacu",
                "description": "Azuriranje zeljenih podataka o zeljenom vozacu",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID vozaca",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "ime_vozaca",
                        "in": "body",
                        "description": "Ime vozaca",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "prezime_vozaca",
                        "in": "body",
                        "description": "Prezime vozaca",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podaci azurirani"
                    },
                    "400": {
                        "description": "Nedostaju podaci o vozacu ili je ID netocan"
                    }
                }
            },
            "delete": {
                "summary": "Brisanje odabranog vozaca",
                "description": "Brisanje vozaca po zeljenom ID-u",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID vozaca",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }],
                "responses": {
                    "200": {
                        "description": "Vozac izbrisan"
                    },
                    "404": {
                        "description": "Ne postoji vozac sa tim ID-om"
                    }
                }
            }

        }
    }
}